requirements.txt
asgiref==3.10.0
bleach==6.2.0
Django==5.2.7
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-summernote==0.8.20.0
django-tinymce==5.0.0
mysqlclient==2.2.7
pillow==12.0.0
pymssql==2.3.9
PyMySQL==1.1.2
pyodbc==5.3.0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.3
transliterate==1.10.2
tzdata==2025.2
webencodings==0.5.1
